+++ R/skip.R
@@ -29,6 +29,9 @@
 #'#' `skip_on_ci()` skips tests on continuous integration systems by inspecting
 #' the `CI` environment variable.
 #'
+#' `skip_on_covr()` skips tests when covr is running by inspecting the
+#' `R_COVR` environment variable
+#'
 #' `skip_on_bioc()` skips tests on Bioconductor by inspecting the
 #' `BBS_HOME` environment variable.
 #'
@@ -175,6 +178,16 @@ skip_on_ci <- function() {
   }
 
   skip("On CI")
+}
+
+#' @export
+#' @rdname skip
+skip_on_covr <- function() {
+  if (!identical(Sys.getenv("R_COVR"), "true")) {
+    return(invisible(TRUE))
+  }
+
+  skip("On covr")
 }
 
 #' @export